Home template uploaded to work with localization
Updates the home template to work with localization.

Note: we initially thought on using the same phrase as key, kinda
forgot that laravel needs to put the file name on front so kinda
defeats it.

Anyway, since generating the translation files is going to be pulled
from the google doc via gulp it’s no problem since you just need to add
the english keys and it will work.