Homepage and profile improvements
Disabled photos/logos from public profile view. Profile doesn’t show
any photo if user hasn’t put one, but theres now a placeholder image in
case we want to enable it.

When logged in, if student home redirects to profile or edit profile
depending if profile has enough information or not.

If the user is a company we redirect them to search

On login now we redirect to home which does the thing mentioned up there