Password change working, db seeder fixes, adds profile visibility option.
You expected bb1b53ecd79db3a8d1903a22e00e3458cbc80eef to be the last
commit of the day, but it was me DIO!

Implements password change. Which should work on all types of user
types.

Adds profile visibilty options to both profiles.

Database seeder for company now fills all  the new fields. And fixes an
issue with student with a now missing column

Fixes an issue that made company activities to not be translated

Edit company profile is now showing the current data.