Change professional skills to personal skills, New validation engine, works for company except for the desired skills logic, should also work for students
This is part of the new validation engine, that updates user data. The
only missing parts are, Desired skills (for company)

The validation engine should work with small template tweaks for the
student user common info and with more tweaks with the detailed info.

The data validation engine will be derived from this too.

Also fixed company to work with personal skills instead of
professional. And added a few convenience methods for accessing some
data more easily.